+++ example/annotation/app.js
@@ -216,7 +216,7 @@ function editAnnotation(region) {
     var form = document.forms.edit;
     form.style.opacity = 1;
     (form.elements.start.value = Math.round(region.start * 10) / 10),
-        (form.elements.end.value = Math.round(region.end * 10) / 10);
+    (form.elements.end.value = Math.round(region.end * 10) / 10);
     form.elements.note.value = region.data.note || '';
     form.onsubmit = function(e) {
         e.preventDefault();