+++ CHANGES.md
@@ -1,9 +1,67 @@
 wavesurfer.js changelog
 =======================
 
-4.0.0 (unreleased)
+x.x.x (unreleased)
 ------------------
 
+- Use Webpack 5 for build (#2093)
+
+4.3.0 (12.12.2020)
+------------------
+
+- Add `relativeNormalization` option to maintain proportionality between
+  waveforms when `splitChannels` and `normalize` are `true` (#2108)
+- WebAudio backend: set playback rate modifying directly the playback
+  property of the source node (#2118)
+- Spectrogram plugin: Use `ImageData` to draw pixel-by-pixel (#2127)
+
+4.2.0 (20.10.2020)
+------------------
+
+- Fix performance issues with `seekTo` while audio is playing (#2045)
+- Trigger `waveform-ready` event when provided peaks are drawn (#2031)
+
+4.1.1 (24.09.2020)
+------------------
+
+- Revert Code cleanup for Observer class (#2069)
+
+4.1.0 (16.09.2020)
+------------------
+
+- Don't call HTMLMediaElement#load when given peaks and preload == 'none'.
+  Prevents browsers from pre-fetching audio (#1969, #1990)
+- `seekTo` bugfix inc. basic unit tests (#2047)
+- Fix unhandled `AbortError` thrown during `cancelAjax` (#2063)
+- Remove `util.extend`: deprecated since v3.3.0 (#1995)
+- Remove `util.ajax`: deprecated since v3.0.0 (#2033)
+- Regions plugin:
+  - Removed `col-resize` cursor when resize is disabled (#1985)
+  - Improved and unified loop playback logic (#1868)
+  - Check `minLength` before resizing region (#2001)
+  - Dragging and resizing will continue outside canvas (#2006)
+  - `regionsMinLength` parameter to assign a min length to those regions for which the `minLength`  is not specified (#2009)
+  - Revert PR #1926 click propagation on regions. Use event parameter passed
+    in `region-click` if you need `stopPropagation`. (#2024)
+  - Edgescroll works for both edges (#2011)
+- Microphone plugin: move to separate directory (#1997)
+- Minimap plugin: move plugin to separate directory (#1999)
+- Cursor plugin: move plugin to separate directory (#1998)
+- Elan plugin: move plugin to separate directory (#2019)
+- Spectrogram plugin: move to separate directory (#1996)
+- Mediasession plugin: move to separate directory (#2020)
+- Timeline plugin: move to separate directory (#2018)
+
+4.0.1 (23.06.2020)
+------------------
+
+- Fixes for event handling with certain plugins (regions, microphone).
+  The crash would have involved '_disabledEventEmissions' (#1975)
+
+4.0.0 (21.06.2020)
+------------------
+
+- Fixed mediaelement-webaudio playback under Safari (#1964)
 - Fixed the `destroy` method of the `MediaElementWebAudio` backend. Instead of
   destroying only the media element, the audio nodes are disconnected and the
   audio context is closed. This was done by splitting the `destroy` method of the
@@ -29,6 +87,7 @@ wavesurfer.js changelog
 - Added checks in `minimap` plugin for `drawer` presence (#1953)
 - Add `setDisabledEventEmissions` method to optionally disable calls to event handlers for specific events (#1960)
 - Drawer: removed private methods to allow overriding them (#1962)
+- Add optional `setMute` method to backends to fix muting behavior with the `MediaElement` backend (#1966)
 
 3.3.3 (16.04.2020)
 ------------------